refactor(tui): split ui/dialogs.rs into dialogs/ submodule by domain
