+++ crates/kanban-tui/src/ui/dialogs/mod.rs
@@ -1,0 +1,9 @@
+mod boards;
+mod cards;
+mod columns;
+mod sprints;
+
+pub(super) use boards::*;
+pub(super) use cards::*;
+pub(super) use columns::*;
+pub(super) use sprints::*;
